# backend/routes/profiles.js
const express = require('express');
const { toCamelCase } = require('../utils');
const { getProfileId } = require('../middleware/profile');
const { asyncHandler } = require('../lib/errors');
module.exports = function ({
  apiRateLimiter,
  logError,
  seedThreeTierProfiles,
  requireAuth,
  demoProfileIds = [1, 2, 3],
}) {
  const router = express.Router();

  function updateProfileHandler(req, res) {
    if (!req.session.userId) {
      return res.status(401).json({ error: 'Unauthorized' });
    }
    const pid = parseInt(req.params.id);
    const { name } = req.body;
    const profile = req.repos.profiles.getById(pid);
    if (!profile) return res.status(404).json({ error: 'Profile not found' });
    if (profile.user_id !== req.session.userId) {
      return res.status(403).json({ error: "Cannot modify another user's profile" });
    }
    if (name !== undefined) {
      if (!name.trim()) return res.status(400).json({ error: 'Name is required' });
      req.repos.profiles.updateName(pid, name.trim());
    }
    const updated = req.repos.profiles.getById(pid);
    res.json(updated);
  }

  router.get(
    '/api/profiles',
    requireAuth,
    apiRateLimiter,
    asyncHandler((req, res) => {
      let profiles;
      if (req.session.userId) {
        profiles = req.repos.profiles.listByUserId(req.session.userId);
        if (profiles.length === 0) {
          req.repos.profiles.create('Default', req.session.userId);
          profiles = req.repos.profiles.listByUserId(req.session.userId);
        }
      } else {
        profiles = req.repos.profiles.allByIds([1, 2, 3]);
      }
      const txCounts = req.repos.transactions.countsByProfile('transactions');
      const acctCounts = req.repos.accounts.countsByProfile('accounts');
      const budgetCounts = req.repos.budgets.countsByProfile('budgets');
      const result = profiles.map((p) => ({
        ...p,
        transaction_count: txCounts[p.id] || 0,
        account_count: acctCounts[p.id] || 0,
        budget_count: budgetCounts[p.id] || 0,
      }));
      res.json(result);
    })
  );

  router.post(
    '/api/profiles',
    apiRateLimiter,
    asyncHandler((req, res) => {
      if (!req.session.userId) {
        return res.status(401).json({ error: 'Unauthorized' });
      }
      const { name } = req.body;
      if (!name || !name.trim()) return res.status(400).json({ error: 'Name is required' });
      const existing = req.repos.profiles.getByName(name.trim());
      if (existing)
        return res.status(400).json({ error: 'A profile with this name already exists' });
      const result = req.repos.profiles.create(name.trim(), req.session.userId);
      res.json({
        id: result.lastInsertRowid,
        name: name.trim(),
        transaction_count: 0,
        account_count: 0,
        budget_count: 0,
      });
    })
  );

  router.put('/api/profiles/:id', apiRateLimiter, updateProfileHandler);
  router.patch('/api/profiles/:id', apiRateLimiter, updateProfileHandler);

  router.delete(
    '/api/profiles/:id',
    apiRateLimiter,
    asyncHandler((req, res) => {
      if (!req.session.userId) {
        return res.status(401).json({ error: 'Unauthorized' });
      }
      const pid = parseInt(req.params.id);
      if (pid === 1) return res.status(400).json({ error: 'Cannot delete the default profile' });
      const profile = req.repos.profiles.getById(pid);
      if (!profile) return res.status(404).json({ error: 'Profile not found' });
      if (profile.user_id !== req.session.userId) {
        return res.status(403).json({ error: "Cannot delete another user's profile" });
      }
      if (req.repos.profiles.profileCount() <= 1)
        return res.status(400).json({ error: 'Cannot delete the last profile' });
      req.repos.profiles.deleteAllDataForProfile(pid);
      res.json(toCamelCase({ ok: true }));
    })
  );

  router.post(
    '/api/profiles/reseed-demo',
    apiRateLimiter,
    asyncHandler((req, res) => {
      if (!req.session.userId) {
        return res.status(401).json({ error: 'Unauthorized' });
      }
      // Clear every demo profile, not just the active one: seedThreeTierProfiles() rebuilds all
      // three, and it skips any profile that still has transactions. Clearing one and seeding
      // three meant "restore the three example profiles" restored whichever profile happened to
      // be active and left the other two as they were — and a profile that had been emptied some
      // other way got a SECOND copy of its accounts, holdings and goals stacked on top. Scoped to
      // the profiles this user owns, so the endpoint can never wipe another account's data.
      //
      // Clear EVERY table too, not the hand-picked few this used to name: the seeder inserts
      // holdings and recurring rows unconditionally, so anything left behind comes back doubled.
      const owned = req.repos.profiles
        .allByIds(demoProfileIds)
        .filter((p) => p.user_id === req.session.userId);
      if (owned.length === 0) {
        return res.status(403).json({ error: 'The demo profiles belong to another account' });
      }
      for (const profile of owned) req.repos.profiles.clearDataForProfile(profile.id);
      seedThreeTierProfiles();
      res.json({ ok: true, message: 'Demo data has been restored' });
    })
  );

  router.delete(
    '/api/profile/data',
    apiRateLimiter,
    asyncHandler((req, res) => {
      const pid = getProfileId(req);
      // "Delete all my data" means all of it — holdings, tags, receipts and housing included.
      req.repos.profiles.clearDataForProfile(pid);
      const defaults = [
        { name: 'Housing', color: '#ef4444', icon: 'home', type: 'expense', tax_deductible: 0 },
        {
          name: 'Food & Dining',
          color: '#f97316',
          icon: 'utensils',
          type: 'expense',
          tax_deductible: 0,
        },
        {
          name: 'Transportation',
          color: '#eab308',
          icon: 'car',
          type: 'expense',
          tax_deductible: 0,
        },
        { name: 'Healthcare', color: '#22c55e', icon: 'heart', type: 'expense', tax_deductible: 0 },
        {
          name: 'Entertainment',
          color: '#06b6d4',
          icon: 'film',
          type: 'expense',
          tax_deductible: 0,
        },
        {
          name: 'Shopping',
          color: '#8b5cf6',
          icon: 'shopping-bag',
          type: 'expense',
          tax_deductible: 0,
        },
        { name: 'Utilities', color: '#64748b', icon: 'zap', type: 'expense', tax_deductible: 0 },
        { name: 'Education', color: '#ec4899', icon: 'book', type: 'expense', tax_deductible: 0 },
        { name: 'Salary', color: '#22c55e', icon: 'briefcase', type: 'income', tax_deductible: 0 },
      ];
      req.repos.categories.seedDefaults(pid, defaults);
      res.json(toCamelCase({ ok: true }));
    })
  );

  router.put(
    '/api/profiles/me/password',
    apiRateLimiter,
    asyncHandler((req, res) => {
      if (!req.session.userId) {
        return res.status(401).json({ error: 'Unauthorized' });
      }
      const { currentPassword, newPassword } = req.body;
      if (!currentPassword || !newPassword) {
        return res.status(400).json({ error: 'Current and new password are required' });
      }
      const bcrypt = require('bcrypt');
      if (newPassword.length < 8)
        return res.status(400).json({ error: 'Password must be at least 8 characters' });
      if (!/[A-Z]/.test(newPassword))
        return res.status(400).json({ error: 'Password must contain an uppercase letter' });
      if (!/[a-z]/.test(newPassword))
        return res.status(400).json({ error: 'Password must contain a lowercase letter' });
      if (!/[0-9]/.test(newPassword))
        return res.status(400).json({ error: 'Password must contain a number' });
      if (!/[^A-Za-z0-9]/.test(newPassword))
        return res.status(400).json({ error: 'Password must contain a special character' });

      const user = req.repos.users.getById(req.session.userId);
      if (!user) return res.status(404).json({ error: 'User not found' });
      const valid = bcrypt.compareSync(currentPassword, user.password_hash);
      if (!valid) return res.status(400).json({ error: 'Current password is incorrect' });
      const hash = bcrypt.hashSync(newPassword, 12);
      req.repos.users.updatePassword(req.session.userId, hash);
      res.json({ ok: true, message: 'Password changed successfully' });
    })
  );

  return router;
};