# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Finance Manager API",
    "description": "REST API for the Finance Manager personal finance tracker. All endpoints are profile-scoped via the `X-Profile-Id` header.",
    "version": "1.1.0",
    "license": { "name": "AGPL-3.0-only" },
    "contact": { "url": "https://github.com/Komediruzecki/finance-manager" }
  },
  "servers": [{ "url": "http://localhost:3847", "description": "Local development" }],
  "security": [{ "cookieAuth": [] }],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    },
...
  },
  "paths": {
...
    "/api/profiles": {
      "get": {
        "tags": ["Profiles"],
        "summary": "List all profiles",
        "responses": { "200": { "description": "Array of profiles with counts" } }
      },
      "post": {
        "tags": ["Profiles"],
        "summary": "Create profile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": { "name": { "type": "string" } }
              }
            }
          }
        },
        "responses": { "201": { "description": "Profile created" } }
      }
    },
    "/api/profiles/{id}": {
      "put": {
        "tags": ["Profiles"],
        "summary": "Update profile",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "name": { "type": "string" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "Profile updated" } }
      },
      "patch": {
        "tags": ["Profiles"],
        "summary": "Patch profile",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Profile patched" } }
      },
      "delete": {
        "tags": ["Profiles"],
        "summary": "Delete profile",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Profile deleted" } }
      }
    },
    "/api/profiles/reseed-demo": {
      "post": {
        "tags": ["Profiles"],
        "summary": "Reseed demo data",
        "description": "Clears every demo profile the authenticated user owns and rebuilds all three example profiles (Low/Mid/High Income) as a set. The X-Profile-Id header is not used.",
        "responses": {
          "200": { "description": "Demo data reseeded" },
          "403": { "description": "The demo profiles belong to another account" }
        }
      }
    },
...
  },
...
}